Week07.Buffer

Documentation

class Buffer b where

Methods

toString :: b -> String

Convert a buffer to a String.

fromString :: String -> b

Create a buffer from a String.

line :: Int -> b -> Maybe String

Extract the nth line (0-indexed) from a buffer. Return Nothing for out-of-bounds indices.

replaceLine :: Int -> String -> b -> b

replaceLine n ln buf returns a modified version of buf, with the nth line replaced by ln. If the index is out-of-bounds, the buffer should be returned unmodified.

numLines :: b -> Int

Compute the number of lines in the buffer.

value :: b -> Int

Compute the value of the buffer, i.e. the amount someone would be paid for publishing the contents of the buffer.

safeIndex :: Int -> [a] -> Maybe a